Avoid double-boxing and using trait objects by making trampoline generic
This makes calls cheaper (at the extreme, boxing a non-capturing closure
or a fn, which are zero-sized, is a no-op) at the cost of monomorphizing
the trampoline and destructor for each closure type.
The zero-sized optimization required also getting rid of RefCell but it
wasn't really expected to ever blow up and AFAIK never did.